Se agrega a la tabla el importe total del pedido

+++ FoodApp.DataLayer/Extensions/MapExtensions.cs
@@ -80,7 +80,7 @@
             {
                 OrderIdentifier = reader.Get<long>("OrderIdentifier"),
                 ItemsCount = reader.Get<int>("ItemsCount"),
-                ItemsTotalAmount = reader.Get<int>("ItemsTotalAmount"),
+                ItemsTotalAmount = reader.Get<decimal>("ItemsTotalAmount"),
                 Customer = reader.ToCustomer(),
                 OrderDetailList = reader.Select(r => r.ToOrderDetail()).ToList()
             };
